requirements.txt
# Available in jessie-backports
Django==1.8.14

# Available in Debian
psycopg2
PyYAML

# From the cheeseshop
python-memcached
wafer==0.4.1
